Make maintainers available to all packages
This change applies an overlay to the pkgs instance in flake.nix to
extend the lib.maintainers set with the custom maintainers defined in
lib/maintainers.nix. This makes the custom maintainers available to all
packages in the repository by default, so they don't need to be passed
explicitly to each package.

This fixes the build error where the ressu maintainer was not found.